Docker Compose deployment tweaks
This change aims to make it easier to build/deploy the proxy
via Docker Compose from a clone of this repo.

A .dockerignore file was created to reduce image size
and improve the layer cache hit rate.

This change also restricts external access to the web proxy port
to the host machine by default. To revert to the old behavior,
add the following environment variable:
```
SCREEPS_PROXY_HOST=0.0.0.0
```

I also updated the docs with instructions on how to deploy
on a machine that doesn't have Steam installed.
`vendor/*.nw` was added to .gitignore/.dockerignore
to prevent anybody using this approach from accidentally
distributing the Screeps client package.