feat: Add CLI flags for external client access when hosted (#46)
This PR updates the client app to allow for specifying the external
address, port, and protocol of the client, for use with the server
listings.

This allows the client to be used from within a container, or from
behind a reverse proxy, where the `host` and `port` parameters used for
local binding/listening do not match the hostname and port used to
access the client.

Behavior of the client should match the existing behavior when the new
CLI arguments are not used.